requirements.txt
alabaster==0.7.12
atomicwrites==1.3.0
attrs==19.3.0
Babel==2.8.0
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
coverage==5.1
docutils==0.16
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.13.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-cloud-bigquery==1.24.0
google-cloud-core==1.3.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
httplib2==0.17.2
idna==2.9
imagesize==1.2.0
importlib-metadata==1.6.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mock==4.0.2
more-itertools==8.2.0
nose==1.3.7
oauthlib==3.1.0
packaging==20.3
pbr==5.4.5
pluggy==0.13.1
protobuf==3.11.3
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
Pygments==2.6.1
pyparsing==2.4.7
pytest==5.4.1
pytest-cov==2.8.1
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
snowballstemmer==2.0.0
Sphinx==3.0.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
uritemplate==3.0.1
urllib3==1.25.8
wcwidth==0.1.9
zipp==3.1.0
